+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -5,7 +5,8 @@ Release History
 
 unreleased
 ++++++++++++++++++
-Fix --resource-type parsing for the lock command to accept <resource-namespace>/<resource-type>
+* Fix --resource-type parsing for the lock command to accept <resource-namespace>/<resource-type>
+* Add support for specifying deployment parameters using KEY=VALUE syntax.
 
 2.0.7 (2017-05-30)
 ++++++++++++++++++